Implement `hanoverd builder`
This can be used to run hanoverd in a builder mode.

In this mode, when hanoverd sees a new git revision, it builds, pushes
to a specified repository, then notifies hookbot.

Hanoverd servers can then subsequently listen on that hookbot URL to
hear of new images that should be fetched.