Force-delete duplicate containers
At the moment, if a container with our name already exists, it
fails to come up. This can happen if, for example, the machine
reboots and the previous one fails to get removed correctly.

It's not my preferred solution since it means deleting something, but
crashing is worse right now.