Label containers rather than naming them
This fixes #50, again. Docker's bindings changed the way it propagated
errors, causing the delete-and-retry path to break. Instead we just do
away with names and instead label them. This means we might end up
with some garbage lying around which needs cleaning. In the long run,
`docker run --rm` would hopefully do the right thing. In the short run
we need a garbage collector for images anyway, so there might as well be
one for containers too.