Fix container superceding
This matters for containers which take a while to start up, where
another request may come in before the first completes (such as on the
CSPS backend during testing). It's important to kill the old one before
trying to start the next one otherwise it may run out of memory.