Detect server non-response during hand over
The integration test now notices if we don't get a valid response from
the server during a handover.

We expect that if we ever saw a response from a server, we must continue
to see responses from servers during handovers, otherwise something has
gone wrong.

If there is a failure, we get a blank response from simpleGetHTTP.

I have tested this by modifying the mulitple-ports program to
return a blank response when the hostname reaches hanoverd_10.