Supply empty authentication during push
Docker was returning:

```
Bad parameters and missing X-Registry-Auth: EOF
```

This was coming from
[here](https://github.com/docker/docker/blame/829ea91bd16c006facf2948cab89302ef2db7306/api/server/router/image/image_routes.go#L151).

It looks like you can't get away without supplying an X-Registry-Auth
header, even though a blank one works. I chose to supply an empty json
dictionary as authentication since it gets unserialized as json into an
[types.AuthConfig](https://github.com/docker/docker/blob/829ea91bd16c006facf2948cab89302ef2db7306/api/types/auth.go).

I don't know how this ever worked, or how I came to convince myself this
was working in the past, it looks like these code paths have been the
same for years.

I have tested this on QA and it appears to work. We'll need to cut a new
hanoverd release and do an infrastructure update.